Clean up embedded code and init
